requirements.txt
openai
langchain
langchain_community
streamlit
huggingface_hub
langchain_openai
python-dotenv